Add pending state to gman
This commit adds a pending state to the GMan head resource.
When a task is passed from a FaaS controller to a FaaS executor
it should perform the following:
[controller]
1. Request new taskID with status started
2. pass execution to executor
3. PUT status of taskID with "delegated"
4. PUT status of taskID with "completed"
[executor]
1. request new taskID with status received
2. perform work
3. PUT status of taskID with "completed"

However the previous logic introduced a timing issue where a task
would be marked as "completed" when being passed from the controller
to the executor BEFORE the executor could request a new task with
"received", which would put the thread back into a running state.

By introducing a "pending" task state we can transition a task
from running to "pending" when a task event with status "delegated"
comes into GMan. When a status of "received" is seen we will remove
the task from the "pending" list and append the task to the "running"
list.

By using the task object instead of a simple counter to track
delegations <-> receives we can add additional logic at a later
point to return which task is "stuck" in a pending state if the
"pending" list never empties.