Fix variable name collision
When a task event with status "received" hits gman
it will enter the for loop which assigns a loop variable of "pending"
which will override the pending task state list that is being checked
later on in the function.

Changing the loop variable "pending" to "pending_task" resolves
the issue.